Try to fix our "don't rerun autoconf" code in qthread/hwloc Makefiles
Since we store our our third-party libraries untarred they lose the timestamps
they had when they were tarballs. For packages that use autoconf/configure, we
try to touch some files in a certain order to prevent autoconf from rerunning,
but we never seem to get this quite right.

A user ran into another case where it seems like we're touching files too fast
for the filesystem's timestamp resolution. This just adds a sleep before we
touch configure to ensure it's newer than the m4 files. This partially reverts
some changes in #69 and #4504