Fix casts, enum first function, and size for enums in generic functions
This was like playing whack-a-mole.  Some of the enum generated functions are
placed in the same scope as the enum definition (like assignment and the
enumerate function), while others are inserted at upper scopes such as
ChapelBase.  The ones that are inserted at the same scope as the enum definition
get automatically updated when we copy the function, but the other ones don't
get checked and so must be regenerated.  Where I was originally trying to create
them ran into problems because the symbol for the new EnumType hadn't been
updated to the new TypeSymbol, and generating them after the new TypeSymbol was
created ran into issues because it hadn't been inserted in the tree yet.  So,
generate them after the DefExpr for the TypeSymbol had been copied entirely.

It's not ideal, but I think the ideal solution would be unreasonable at this
time.  I'd really like more of these functions to be made methods on the type
so that they could get updated with the type like we do for AggregateTypes, but
that would require a little more discussion.