Clean up code by removing support for LLVM older than 4 (#7643)
Clean up code by removing support for LLVM older than 4 (#7643)

Supporting old versions of LLVM caused considerable clutter (over 500 lines worth). This change removes the old #ifdefs. Where reasonable, it also further consolidates the code once the conditionals are removed.

This implements part of #6993 .

While testing this, I ran up against ClangInfo being declared as a struct in one place and a class in another. To get it to compile with the latest Mac Clang, I had to make it consistently a struct.