Fix issue with casting an enum's constants to certain types
Resolves #5412

If an enum constant's value was cast to int(*), uint(*), etc, we would run into
an internal error at EnumType::sizeAndNormalize() due to expecting to get a
SymExpr for the constant's value when it was another type of Expr (in this case,
a CallExpr).  This did not occur for int, uint, etc, nor for type aliases.  It
turns out the source of the problem was the iteration through the initialization
expression for the constant during resolveTypeOrParamExpr(), which was
terminating early due to replacing some of the nested contents of the
initialization expression (meaning the currently examined expression was no
longer in the tree, so e->parentExpr was NULL).  This means that if we had used
a complicated expression, we might have gotten the wrong type back in certain
circumstances as well.

This simple change fixes the early termination of the loop, allowing the enum
constant expression to get completely resolved and not trigger the internal
error.

Fixes test/types/enum/typed.chpl