Fix CHPL_COMM_DEBUG=1 OPTIMIZE=1 builds
CHPL_COMM_DEBUG builds gasnet with --enable-debug and GASNet has some checks to
ensure that --enable-debug isn't combined with -O3 or -DNDEBUG. We have force
thrown -O0 for CHPL_COMM_DEBUG for a while, but we've been mistakenly throwing
-DNDEBUG since #1099.

This just throws -UNDEBUG if CHPL_COMM_DEBUG is set. Note that if somebody is
building with CHPL_COMM_DEBUG, they might want to build with OPTIMIZE=0 so the
rest of the runtime is built with debugging enabled, but there's no reason you
shouldn't be able to just enable debugging for gasnet.