Fix enum constants that depend on previous ones
This commit adds support for things like:

 enum x {
   a,
   b,
   c = a + b
 }

The idea here is that as long as the uses of the other
enum elements are handled in order, we can support this.
The implementation amounts to assigning the enum values
immediates in a single loop that goes through resolving
the enum values. This logic moves from EnumType::sizeAndNormalize()
to ensureEnumTypeResolved().

Besides making that change, fixed several more enum bugs
related to enums containing large integers.