Adjust intChoices.good for different enum behavior
Consider the following program:

  enum myenum { X=1, Y=131072 }

  proc foo(x:int(8)) {
    writeln("int(8) version");
  }

  proc foo(x:int(32)) {
    writeln("int(32) version");
  }

  foo(myenum.X);

The numeric type associated with myenum will be uint(32), but
the enum has a min constant of 1 and a max constant of 2**17.
Thus canParamCoerce can return 'true' for converting myenum
to int(32) and that shouldn't be considered a narrowing conversion.
Thus the int(32) version of foo is a better match than the int(8)
version, even though myenum.X could pass to an int(8) argument.