Retire clonedForall future
This test was created to track duplication of loop bodies out of concern
for compilation efficiency. The duplication was caused by branches like
the following in our internal modules:

if numChunks == 1 {
  // serial loop
} else {
  // coforall loop
}

@ronawho suggested that LCALS would be a good benchmark for compilation
time w.r.t loops. I compared compilation time with and without the
serial loop and didn not see a difference. This is a curious result, but
the generated code did indeed contain only one instance of the loop
body.

Since this test checks for something we cannot demonstrate to be
valuable, it will be retired.