Add ability for 1:int(16) to dispatch to uint(8), say
It seems to me that a param value such as 1:int(16) should
still be able to dispatch to a smaller type, such as uint(8).
In my previous commit ("Track enum type min/max"), I changed
fits_in_int/fits_in_uint to no longer require that the immediate
in question be a default-sized int. That had a follow-on effect
that caused several testing failures.

There is a tension between what we'd expect a binary operator (say +)
to do with these sets of arguments:

 1:uint(8) + 1

vs

 1 + 1

This commit addresses this tension by adding 2 more levels
of param-preferred-ness.

 * first level causes it to:
     prefer param vs not param
     prefer conversions that don't narrow a specified type
        (e.g. 1:int(16) to 1:uint(8))

 * second level causes it to:
     prefer an exact type match
     prefer certain conversions to a 3rd type
       (e.g. enum->int prefers default sized int)

 * third level causes it to:
     prefer 'int' param is passed to 'int'
     prefer narrower arguments
     prefer 'int' over 'uint'