With mem=jemalloc, comm=ugni, and hugepages, dynamically extend heap.
In the past we have only provided a fixed-size registered heap when
CHPL_COMM=ugni.  Here, build on the dynamic registration capability
recently added for array allocations, to support dynamic extension of
the main heap also, even when it is composed of registered memory.

The specific file changes are as follows.

mem-jemalloc.c:
Support either a fixed or a dynamically extendable heap.  Localize heap
extension chunks to the NUMA domain we are running on.  Heap extension
chunks belong to specific arenas and jemalloc tries to maintain affinity
between threads and the arenas they allocate from, so arenas will tend
to have chunks with the same NUMA locality.

comm-ugni.c

Only set up an initial heap if the user set the CHPL_RT_MAX_HEAP_SIZE to
indicate they want a fixed one.  Otherwise, support a dynamic heap by
allowing extension chunks to be allocated and registered through the
chpl_comm_memReg*() functions.  Record memory regions during "early"
initialization, before the comm layer is fully set up, but don't
register them until we can do so and broadcast the resulting MDHs.  (We
need to be able to do this because the runtime memory layer has to be
initialized very early in order to support memory allocation by other
layers, including the comm layer itself.)  Don't let tasking layers
multitask on the pthread while holding the mutex that protects memory
registration table updates, or we can end up trying to lock the mutex
recursively.

While here, also bump the number of registered memory regions from 1000
to 1024, which seems more in tune with the other power-of-2 values in
the ugni comm layer.

comm-ugni-launch.c:
Set the 'lg_chunk' component value in the environment variable
CHPL_JE_MALLOC_CONF to indicate the heap extension chunk size when
dynamic heap extension is to be done.  This has to be done in the
launcher before the user program is started, because in the user program
jemalloc will read the CHPL_JE_MALLOC_CONF value in a constructor
(.init) function, before main() is entered, and in the user program we
don't have a way to ensure it's set before that.

comm-ugni-heap-pages.h:
comm-ugni-heap-pages.c:
These are new files, defining chpl_comm_ugni_getSysPageSize() which
returns the system page size and chpl_comm_ugni_getHeapPageSize() which
returns the ugni-specific heap page size.  These are present so that
the launcher and user program definitely use the same values for those
two.  (chplsys.c:chpl_getSysPageSize() could also provide the system
page size, but that file:function is not available in the launcher.)

Makefile.share:
Add comm-ugni-heap-pages.c.

chpl-comm.h:
While here, clarify what chpl_comm_regMemAllocThreshold() returns if the
comm layer cannot or will not allocate memory.