Retire 'useDummy' future
This future was created to track a simpler form of a bug captured in
this test:

test/modules/vass/SSCA2-with-Private-1.chpl

That test by Vass was un-futurized in a05dbed2, where Vass mentions he
could not re-create it in the new world.

Since its creation, this 'useDummy' test has fallen out of sync with
assumptions in other distributions such that it now fails to compile in
a totally different way than was originally intended. It is a nontrivial
effort to get this test to compile again (unrelated to the original
bug), so it will be retired.