Remove traditional CHPL_COMM_USE_{G,LL}DB change from GASNet sources.
For years, each time we've upgraded GASNet we've made a change to their
sources to support the CHPL_COMM_USE_{G,LL}DB environment variable which
says to start a debugger on each of the compute node instances of a
Chapel program that uses comm=gasnet and the amudprun launcher (with
either local or ssh-based spawning).  Now we're going to support this
capability entirely within Chapel itself, so undo the change in GASNet.