Adjust dispatchenums.good for different enum behavior
The reason is similar to, but not quite the same, as the previous commit
adjusting intChoices.good.

The following program shows the behavior:

  enum myenum { X=1, Y=1000 }

  proc foo(x:int(8)) {
    writeln("int(8) version");
  }

  proc foo(x:int(32)) {
    writeln("int(32) version");
  }

  foo(myenum.X);

The numeric type associated with myenum will be uint(16), which fits in to an
int(32). As with the intChoices.good change, the new rules prefer conversions
than aren't "narrowing" (i.e. relying on the enum's value). So in this case,
the int(32) version is preferred because not all enum values would fit in to
int(8).