Make the workaround in test_override_operators.chpl more robust
In commit ea3297cfaf9689e6930c05c28e222d91616cc569, Brad was
improving function resolution for enum constants. At that point,
he pulled part of this test out in to a future and updated this
test to work (somehow). This commit updates it further to work
with my changes. Note that the originally requested functionality
is still documented in the .future file.