Support CHPL_COMM_USE_{G,LL}DB in the Chapel amudprun launcher.
In the past we've supported CHPL_COMM_USE_{G,LL}DB, which starts each
target program instance in its own debugger window, through a mod to the
GASNet amudprun launcher.  That's been removed, and here we add code to
achieve the same effect in our own Chapel amudprun launcher.  If either
of those environment variables is set, then instead of having the GASNet
launcher run the target program directly we have it run an xterm window
for each instance, with the xterms invoking the target program in gdb.

Note that there is still one more step needed.  Currently the two env
vars have the identical effect, which is to run gdb.  But really we need
CHPL_COMM_USE_LLDB to run the (MacOS) lldb debugger instead.  That's
next.