Differentiate between CHPL_COMM_USE_GDB and CHPL_COMM_USE_LLDB.
The first version of this code treated the two environment variables
identically.  Now, differentiate so things work as expected on both
Linux (and other gdb platforms) and MacOS (and other lldb platforms).

While here, add a warning to be printed if we can't find an xterm on
which to base our debugger windows.