Shutdown the memory layer after the comm layer
Previously we were shutting down the memory layer before the comm layer, but
the comm layer should be allowed to allocate/deallocate during exit, so this
was the wrong order. Move mem exit after comm exit, and only run mem exit if
we're doing an orderly "all" exit.

We don't shutdown the tasking layer for an abnormal "any" exit because anything
could still be going on and we just want to shutdown without cleaning up much.
Similarly we don't want to try to shutdown the memory layer, because other
things (like the tasking or comm layer) could still be trying to allocate and
deallocate memory and if the memory layer is shutdown this could just lead to
other confusing errors instead of just exiting.