Clarify code in weakPrefers, avoid non-transitivity
This changes weakPrefers to:
 * prefer bool/enum cast to any int over any uint
 * prefer default-sized int/uint over any other int/uint

This avoids ambiguity say when resolving:

 var x:myenum;
 x + 1;

when the inline proc +(a: int(64), param b: int(64)) overload
in ChapelBase is commented out. It is surprising to me that would
result in ambiguity, and in fact it is due to some non-transitivity
in finding the most specific match.

I find these new rules more satisfying, because they enable
a case like what is shown in test/types/enum/ferguson/enum-uint.chpl:

  enum myenum { v }

  proc foo(x:uint(8)) {
    writeln("uint(8)");
  }

  proc foo(x:uint) {
    writeln("uint");
  }

  var x:myenum;

  foo(x);

It seems a shame to me to generate an ambiguity error in this event,
which is what happens on master before this PR.