Add sub_test filtering for Jira 287
Jira 287 (Clock skew) is a sporadic failure we see mostly on VMs though
occasionally on real hardware too.

This adds sub_test filtering so it's easier to pick out when this error errors.
It happens once in a while in nightly testing (I count 10 occurrences this
year), and this should make it much easier to pick out. Long term we may want
to squash theses errors if they keep occurring since they don't affect
correctness and there's nothing we can do about it, but this at least makes it
easier to identify for now.