Rework runtime environment handling.
Based on review feedback, finally implement long-awaited changes in the
runtime environment handling.  Implement common code to convert env var
values to int64_t and size_t types in addition to the existing chpl_bool
typed conversion, make this code available in both the runtime proper
and the launcher, and rework a number of obvious places where we were
getting the values of environment variables.  (Quite a few others not
related to this PR still remain to be done, however.)