Restructure extendLeaderNew() and its helpers
* Restructure most of the "new" "Part 2" part of the implementation
of forall intents, specifically extendLeaderNew() and its helpers.
It still handles a variety of cases as before, now hopefully in
a more comprehensible manner.

* The following case is now handled nicely and uniformly
for all forall intents. The case is when the parallel iterator
(the one that is invoked by the forall loop)
itself contains a parallel loop (i.e. a forall or a for-loop
that invokes a leader or standalone iterator explicitly).

* The case from the previous bullet where the parallel loop is
a for-loop is handled by converting it to a forall loop first.
Handling a forall loop in this context is much simpler than
handling a for-loop, partly due to a better AST representation.
As a result, `eflopiMap` is no longer needed for the "new" implementation.

* The case when a yield in the parallel iterator is nested
inside 2 (or more) levels of parallel iterators has been causing INT_ASSERT.
Now it generates a "currently not implemeted" error instead.