Clean up setting the jemalloc configuration env var.
More changes reflecting review feedback:

Hardcoding the jemalloc config env var is a bad idea.  There's a slim
chance its basename may change in the future, but greater odds that
we'll end up using a system jemalloc without the chpl_je_ prefix some
day.  So here, in the launcher, make use of our CHPL_JEMALLOC_PREFIX
setting that comes from the runtime side of the build.  First, split out
the preprocessor logic that sets that in the no-op case, in the jemalloc
chpl-mem-impl.h, into its own file.  That gives us a single point of
definition for that prefix info.  Then include that file into the
launcher comm-ugni-launch.c file and adjust the code there to build the
prefixed config var name.  To support this, have the runtime Makefiles
include -D and -I compilation options that will let launcher code being
built for a mem=jemalloc target have access to the runtime include files
and thus definitions for jemalloc.  (Note that we only need access to
the Chapel shim oriented definitions here, we don't need access to the
jemalloc definitions themselves.  Which is good, because the launcher is
actually built to use cstdlib mem, not jemalloc; we don't even have a
build of jemalloc for the launcher.)