Quiet timeouts for privatization test
addAndGetPrivatizedInParallel has been sporadically timing out for cce whitebox
testing ever since that configuration was moved onto a busier machine. It turns
out that concurrent parallel-safe I/O uses a spinlock under qthreads, and this
test was doing a writeln on all cores, which led to timeouts when the machine
was busy. In order to quiet this test in the meantime, this just stores the
results in an array and writes out the array at the end to avoid concurrent IO

For more info see https://github.com/chapel-lang/chapel/issues/7748