Address review feedback in mem=jemalloc shim.
The heap page_size is only needed when localizing new pages, so just get
it then and there instead of storing it in every heap descriptor.  Make
the name of the flag that tells whether we just have a single fixed-size
heap from the comm layer or can extend without bound more descriptive,
and store that outside the heap descriptor as well because it is also
global information.