Replace hugepage_size var with get_hugepage_size() fn.
One more change reflecting feedback review: replace the hugepage_size
variable with a get_hugepage_size() function that returns 0 if we're not
using hugepages, and call that everywhere we referenced the variable
before.  (Reviewer argued that it was hard to prove to oneself that the
variable was always set before it could be read the first time.)  As a
side effect, this restores the page_size local variable removed from
make_registered_heap() (was make_shared_heap()) by an earlier commit for
this same PR, reducing the textual changes for the PR as a whole.