chpl_arch and chpl_compiler support --llvm compiles
When the Chapel compiler runs with --llvm, it uses the compiler
configuration name 'clang-included' to find the runtime. When this
happens, occasionally we also need to work with the original compiler
that would have been used if --llvm were not specified.

This commit adjusts chpl_compiler.get to accept a new llvm_mode
argument which can request:
 * "default" -> use 'clang-included' if we're doing an --llvm build
   (knows "if we're doing an --llvm build" by an environment variable
    because that was the only way I could get it to work consistently and
    also be memoized).
 * "orig" -> get whatever the compiler would be if --llvm were
   not specified
 * "llvm" -> always returns 'clang-included'

Additionally it adjusts chpl_arch to check for prgenv compiles in the
original compiler when doing an --llvm build.