Add printchplenv --llvm
When doing an --llvm compile, the chpl compiler needs different
information from usual with printchplenv. This commit adds a new
command-line option to printchplenv to request this new information.

The new --llvm mode will raise an error unless
chpl_compiler.get('target') returns 'clang-included' which in practice
means that it needs to be run with the environment variable
CHPL_LLVM_CODEGEN set. This is to allow chpl_compiler.get to still be
memoized.

The new variables that will be needed by the chpl compiler are:
 * CHPL_ORIG_TARGET_COMPILER
 * CHPL_MAKE_RUNTIME_SUBDIR
 * CHPL_MAKE_LAUNCHER_SUBDIR