Update preferred system version of LLVM to 4.0 (#7767)
Update preferred system version of LLVM to 4.0 (#7767)

[trivial, not reviewed]

In the combination of #6962 and #7643 , I missed updating the preferred LLVM version from 3.7 to 4.0 in one place. This did not become apparent until we started testing CHPL_LLVM=system on more platforms.

This fixes the last remaining place where the version needs to be bumped.