Fix arguments like type t:object
After working with function resolution, I've realized that it's
significantly more work for the compiler to resolve a 'where' clause than
a type constraint expressed with a : on the argument. Additionally, I
know that many _cast overrides are created and resolved in the process of
normal program compilation, and these typically use where clauses. For
this reason, I was wondering if declaring arguments like 'type t:object'
or 'type t:integral' would work.

So, I tried a few cases and to my surprise most of them were working
already! I only needed to make a minor modification to allow it in the
remaining cases. (In particular, only the type t:someclasstype were not
working).

This pattern does not seem to be used much (if at all) in our code base
currently, so I added a test to exercise it.