Fix registration ordering bug in early setup; clean up.
Performance testing with small hugepages (2mb, e.g.) resulted in
segfaults.  The cause turned out to be that with small hugepages we
could need to add arena extension chunks more than once as a result of
chpl_mem_allocMany() calls coming from the comm layer itself, and it
could work out that the chunk containing mem_regions_map never got
registered.  Here, rearrange so that this doesn't happen.  Mainly, since
we know we'll have to do PUTs to it, get mem_regions_map from memory
that is guaranteed to be registered.

While doing this I also moved the barrier information PMI_Allgather() up
earlier in the code.  It never has really belonged with the one for the
memory regions anyway.  And, I made other changes to standardize some
symbol names associated with the various PMI_Allgather() calls.

While debugging this problem I noticed some flaws in the debugging code,
which I've also fixed.  And, I made it possible to print the memory
regions high water mark without turning on debugging, because at least
until all this memory registration work winds down it may be useful to
see that number without having to rebuild the comm layer with debugging
enabled.

Finally, I also marked all the guaranteed-registered memory allocation
calls as such, and removed the '// * no: not communicated' marker
comments that I'd added quite a while ago in support of the minimum
registration mode.