Add VarSymbol constructor for QualType, QualifiedType::refToRefType
These are helper methods.

VarSymbol accepting a QualifiedType is just more satisfying than
explicitly setting the qualifier after construction.

QualifiedType::refToRefType switches the type to a reference type
(instead of relying solely on the Qualifier). This helps to work with
some parts of the compiler that havn't been updated yet for
QualifiedType.