Don't bother trying to use stack-based remote fork completion flags.
When we need a completion flag for a remote fork, as an optimization
we've been first trying to use one on our stack and only if that fails,
allocating one from the pools in guaranteed-registered memory.  But
that's not really an optimization when there are potentially lots of
registered memory regions, because looking up the memory region for our
stack takes longer than using our special-purpose rf_done allocator.
Here, just always get the remote fork completion flag from the pools.

While here, I also cleaned up the treatment of blocking remote forks in
internal functions for those.