Make our execution limiter race-free
Use filelock (https://pypi.python.org/pypi/filelock) to make our execution
limiter race free in order to avoid timeouts under llvm/whitebox testing.

PR #3928 added an "execution_limiter" to try to prevent multiple executables
from running at the same time. It used a file as a lock in racey manner, but we
didn't think the race would be that big of a deal. However, recently we've seen
timeouts for some whitebox and llvm configurations that are using paratest in
heavily a oversubscribed manner and in all cases it seems to be because
multiple chpl executables are running, meaning the race is hurting us.

Our old implementation used a soft file-based lock, whereas filelock will use a
hard lock when available. filelock is a platform independent lock that uses 1
of 3 platform dependent implementations. It will use `fcntl.flock` on unix
systems (including mac and cygwin) to create a hard lock, `msvcrt.locking` on
windows systems, and a soft file-based lock otherwise. fcntl and msvcrt don't
have issues with 'stale' locks because the lock is held on the file descriptor
and gets dropped by the OS even if the program crashes. The soft file-based
lock can cause stale locks, so if we ever run into a platform where
fcntl/msvcrt aren't supported we should re-add our own implementation that can
deal with stale lock files.

Also note that fcntl/msvcrt will leave the actual file used for locking around.
Removing the file would "unlock" the lock, so there'd be a race between locking
and unlocking/removing (we'd effectively have a double unlock). This is a
common problem for file-based locks. Since we're only using a lockfile per
user, this shouldn't be a big deal, but if we start using the execution limiter
for more things we might need to find a different solution.