Add a future for a weird case with fancy generic record initializers
The record in this test has an initializer that leads to a different type than
what you would expect from just looking at the arguments.  In the case of an
explicit initialization with both a type and "new" call, I expected this case to
either a) take the lhs side type and encounter a type mismatch for giving it the
value from the "new" call which yielded a different type, or b) treat the
explicit type as a guideline and use the type yielded by the "new" call.
Instead, it takes the type of the type declaration and gives no error - which
seems like the worst outcome.