Add a couple of tests demonstrating an issue with arrays and records w/ init()
Currently the creation of an array of records that define initializers, via the
parallel loop expression syntax (var x = [i in myRange] new rec(args);),
requires that the record define a zero argument initializer.  Should this be
necessary?

When this issue is resolved, we will be able to update the tests
test/users/vass/km/array-of-records-crash-2*.chpl to use initializers instead of
constructors