Avoid memory region lookups for things in guaranteed-registered memory.
The memory region maps, fork request buffer free flags, and remote fork
completion flags are in memory that is guaranteed to be registered.
Here, distribute the region descriptors for this memory to all the nodes
during setup, and thus avoid the need to look up the memory descriptor
when we need to remotely PUT to any of those data structures.  The main
change is to add gnr_mreg_map and use PMI_Allgather() to fill it in, so
that every node has the region descriptor for every other node's
guaranteed-registered memory.  Add optional arguments to do_remote_put()
and do_remote_put_V() to allow callers to pass in the target memory
regions when they are known.  And then where we know that we're doing
PUTs to memory region maps, fork request buffer free flags, or remote
fork completion flags, pass the target nodes' guaranteed-registered
memory region descriptors directly instead of looking them up.

Relatedly, produce an explicit message in do_remote_put_V() when a
target address isn't in any memory region known to us and we cannot
proxy, and change chpl_comm_bundleData_t's completion indicator member
name from "done" to "rf_done" to clarify that its use is the same as all
other things named that.

Also while here, count local and remote memory region lookups in the
performance statistics (the remote lookup count being unexpectedly large
was actually the driver for making the main change in this commit); make
the may_proxy_false and may_proxy_true enum values explicitly false and
true, respectively; and change a number of checks whether the target
locale is the current one (which the caller is supposed to handle) into
regular asserts so that they're only done in the debug case.