Make aarch64 atomics the same as others (#7804)
Make aarch64 atomics the same as others (#7804)

[reviewed by @ronawho ]

This change gets rid of a special case and starts treating 64-bit ARM atomics the same as atomics in other environments.

In the early stages of the 64-bit ARM port, CHPL_ATOMICS was forced to cstdlib to accommodate a certain cross compiler. However, that compiler ended up never being fully developed enough to compile Chapel. The compilers that do work for 64-bit ARM are capable of the same atomics settings as everything else.