Ignore import errors for filelock
If filelock can't be imported, just ignore the error. Most of the time the
execution limiter won't be used, so the error will be silently ignored. If the
execution limiter is actually being used this just moves the error from import
time to when it tries to grab the lock, which shouldn't be a big deal.