Resolve issue with sync fields without an initial value and default initializers
Prior to this change, when the type with a default initializer had a sync field
with no explicit initial value and the default initializer was called with no
explicit value for the corresponding argument, we would encounter a deadlock.
This was because the default initializer was trying to readFE the value it
received from the wrapper, but the value received was an empty sync.

This change adds a special case to the transformation of field initialization
in Phase 1, which will only apply if all of the following conditions are met:
1) we are in a compiler generated initializer
2) the type of the field is relatively simple and can be determined to be a
   sync
3) the initExpr being used to initialize the field is a simple access to an
   argument
3a) the type of the argument is relatively simple and can be determined to be a
    sync

It does not yet handle the involvement of type aliases or type functions.  But
it does allow the related future to pass.