Make debugging code better/cheaper
getenv is surprisingly slow, which has the unfortunate impact of killer
task-spawn/lcals performance and improving prk-stencil (because it allows the
tasks to go to sleep)

Use a static int and a setter function instead of using an env var. Also make
it more clear that these are experimental chapel hooks for qthreads and not
really qthread proper api calls.