Added FnSymbol::singleInvocation()
singleInvocation() returns the CallExpr that is the single call
to 'this' FnSymbol. If there is no such call, return NULL.

There are a couple of places in the compiler where I would like
to assert that the call I am looking at is a single invocation.
singleInvocation() allows me to do that cheaply,
thanks to firstSymExpr() and lastSymExpr().

This PR also uses singleInvocation() to add an assertion in
implementForallIntents2.cpp and to replace an expensive check
in iterator.cpp.