Cleanup for addForallIntent(), addTaskIntent()
I switched the parser to produce a ShadowVarSymbol instead of the
struct IntentExpr, when parsing each 'intent_expr" of a where-clause.

This necessitated some cleanup in related code.

The primary shortcoming of the old scheme is that we still needed to
generate a ShadowVarSymbol() for an IntentExpr, if the where-clause
is for a forall loop. By the time we run "new ShadowVarSymbol()",
however, the parser's yystartlineno points to the start of the where-
clause, rather than to the current 'intent_expr'. So I had to patch
the resulting nodes' astlocs manually, see the addForallIntent() change
in #7823.

Parser file updates to be committed separately.