Add BurntSushi test suite with JSON compare PREDIFF
Solutions provided as json, so we use python script (compare.py)
to load the produced json.out file and the solution .json file, and
perform a json diff. The output of the diff is written to the .good
file.

If the .good file is empty, the test succeeded.

The PREDIFF emits a print statement to confirm that it is running the
script with the expected arguments.