Wrappers for functions with FLAG_VOID_NO_RETURN_VALUE should have the same flag
When creating a wrapper for a function with FLAG_VOID_NO_RETURN_VALUE, give
the wrapper the same flag.  This fixes an issue where a function that doesn't
return is assignable to a variable if it is promoted.  It should be an error.

Remove the .future and .bad files for a test about this.