Fix missing cast in KNL locale model (#7886)
Fix missing cast in KNL locale model (#7886)

[trivial, tested but not reviewed]

While putting a new KNL test machine through its paces, a missing cast became apparent in a back corner of the KNL locale model. This change adds the cast (in two places), and supplies a test that used to fail before the fix.