Mass-change CHPL_LLVM!=llvm to CHPL_LLVM==none in SKIPIFs
In SKIPIF and suppressif files, CHPL_LLVM was often tested against
the string "llvm". Worked OK if CHPL_LLVM was a binary, w the only
alternative values were "llvm" or "none".

However, now I am trying to test with CHPL_LLVM = system, and these
changes should make the same test cases fire (or not) when CHPL_LLVM
is "system" as when CHPL_LLVM was "llvm".